moved lastobs outputting to nwm_ouput then to nhd_io for calcuation
